feat: code refactor
