+++ prompts/supervisor.py
